requirements.txt
fastapi
uvicorn
python-dotenv
chromadb
google-generativeai
pydantic